Fix - miiiCasa Bar can't be expanded correctly.
After Hunter's modification for triggering onviewload(),
all onviewload() methods in miiiCasa Bar modules won't be triggered.
The reason is that YUI contentReady event will not be triggered when
the page is inside an <iframe/>.

I add (window !== top) to check if this page is in iframe.
If it's in iframe, using document.getElementById() for detection.
If it's not in iframe, using general contentReady event.